moved Splash component from pages to components directory
