+++ client/src/components/Splash/style.css
